backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pip==24.0
psycopg2-binary==2.9.11
pyzmq==27.1.0
Werkzeug==3.1.4
zmq==0.0.0